Extract user endpoints
