+++ tsconfig.json
@@ -9,6 +9,7 @@
     "jsx": "react",
     "baseUrl": ".",
     "paths": {
+      "@api/*": ["src/api/*"],
       "@app/*": ["src/app/*"],
       "@lib/*": ["lib/*"]
     }
