+++ tsconfig.server.json
@@ -7,6 +7,7 @@
     "jsx": "react",
     "baseUrl": ".",
     "paths": {
+      "@api/*": ["./src/api/*"],
       "@lib/*": ["./lib/*"]
     }
   },
